Renamed BlazorSample to BlazorClientSideSample
