+++ samples/BlazorClientSideSample/Pages/Index.razor
