+++ samples/BlazorClientSideSample/Program.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.AspNetCore.Blazor.Hosting;
+
+namespace BlazorClientSideSample
+{
+    public class Program
+    {
+        public static IWebAssemblyHostBuilder CreateHostBuilder(string[] args) =>
+            BlazorWebAssemblyHost.CreateDefaultBuilder()
+                .UseBlazorStartup<Startup>();
+
+        public static void Main(string[] args)
+        {
+            CreateHostBuilder(args).Build().Run();
+        }
+    }
+}
