+++ samples/BlazorClientSideSample/Shared/MainLayout.razor
