+++ samples/BlazorClientSideSample/Shared/NavMenu.razor
