+++ samples/BlazorClientSideSample/_Imports.razor
@@ -2,8 +2,8 @@
 @using Microsoft.AspNetCore.Components.Layouts
 @using Microsoft.AspNetCore.Components.Routing
 @using Microsoft.JSInterop
-@using BlazorSample
-@using BlazorSample.Shared
+@using BlazorClientSideSample
+@using BlazorClientSideSample.Shared
 
 @using Blazored.Toast
 @using Blazored.Toast.Services
