+++ samples/BlazorClientSideSample/wwwroot/css/open-iconic/font/fonts/open-iconic.woff
