--- samples/BlazorSample/Program.cs
@@ -1,16 +1,0 @@
-﻿using Microsoft.AspNetCore.Blazor.Hosting;
-
-namespace BlazorSample
-{
-    public class Program
-    {
-        public static IWebAssemblyHostBuilder CreateHostBuilder(string[] args) =>
-            BlazorWebAssemblyHost.CreateDefaultBuilder()
-                .UseBlazorStartup<Startup>();
-
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
-    }
-}
