--- samples/BlazorSample/Startup.cs
@@ -1,19 +1,0 @@
-using Blazored.Toast;
-using Microsoft.AspNetCore.Components.Builder;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace BlazorSample
-{
-    public class Startup
-    {
-        public void Configure(IComponentsApplicationBuilder app)
-        {
-            app.AddComponent<App>("app");
-        }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddBlazoredToast();
-        }
-    }
-}
